+++ scripts/similation/simulate_porec_fragments.py
@@ -1,0 +1,15 @@
+#!/usr/bin/env python
+# -*- coding:utf-8 -*-
+
+"""
+
+"""
+
+import argparse
+import logging
+import os
+import os.path as op
+import sys
+
+import cooler
+